requirements.in.txt
pip-tools
requests
openbb
TA-Lib
google-cloud-storage
kaleido
sib-api-v3-sdk
